ber_int_t, ber_tag_t, ber_socket_t, ber_len_t added lber_types.h.nt, lber_types.h.in removal of NULLxxx internal macros (in favor of NULL). ch_free added to slapd,slurpd/ch_malloc.c #define free ch_free (should be removed after s/free/ch_free/g) in proto-slap.h ch_malloc and friends use ber_memalloc and friends

+++ servers/slapd/monitor.c
@@ -129,7 +129,7 @@ monitor_info( Connection *conn, Operation *op )
 	val.bv_len = strlen( buf );
 	attr_merge( e, "totalconnections", vals );
 
-	sprintf( buf, "%d", dtblsize );
+	sprintf( buf, "%ld", (long) dtblsize );
 	val.bv_val = buf;
 	val.bv_len = strlen( buf );
 	attr_merge( e, "dtablesize", vals );
